requirements.txt
Flask==2.2.2
pyttsx3==2.9
SpeechRecognition==3.8.1
wikipedia-api==0.5.4